+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineScaleSetStrategy.cs
@@ -84,7 +84,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             string securityPostureId = null,
             string[] securityPostureExcludeExtension = null,
             bool? enableProxyAgent = null,
-            bool? addProxyAgentExtension = null
+            bool? addProxyAgentExtension = null,
+            string highSpeedInterconnectPlacement = null
             )
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -201,7 +202,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                             AllocationStrategy = skuProfileAllocationStrategy
                         },
                         DoNotRunExtensionsOnOverprovisionedVMs = doNotRunExtensionsOnOverprovisionedVMs ? true : (bool?)null,
-                        OrchestrationMode = orchestrationMode
+                        OrchestrationMode = orchestrationMode,
+                        HighSpeedInterconnectPlacement = highSpeedInterconnectPlacement
                     };
                     if (auxAuthHeader != null)
                     {
@@ -254,7 +256,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             string securityPostureId = null,
             string[] securityPostureExcludeExtension = null,
             bool? enableProxyAgent = null,
-            bool? addProxyAgentExtension = null
+            bool? addProxyAgentExtension = null,
+            string highSpeedInterconnectPlacement = null
             )
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
@@ -353,7 +356,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                             AllocationStrategy = skuProfileAllocationStrategy
                         },
                         DoNotRunExtensionsOnOverprovisionedVMs = doNotRunExtensionsOnOverprovisionedVMs ? true : (bool?)null,
-                        OrchestrationMode = orchestrationMode
+                        OrchestrationMode = orchestrationMode,
+                        HighSpeedInterconnectPlacement = highSpeedInterconnectPlacement
                     };
                     if (auxAuthHeader != null)
                     {