+++ src/Compute/Compute/ChangeLog.md
@@ -24,6 +24,7 @@
 * ComputeRP related cmdlets will now use 2025-04-01 version of the ComputeRP API.
 * Added `-AddProxyAgentExtension` parameter (Switch) to `New-AzVmss` and `New-AzVM`
 * Added `-AddProxyAgentExtension` parameter (Bool) to `Set-AzVMProxyAgentSetting` and  `Set-AzVmssProxyAgentSetting`
+* Added `-HighSpeedInterconnectPlacement` parameter to `New-AzVmssConfig` and `New-AzVmss` cmdlets.
 
 ## Version 10.4.0
 * Added `-InstantAccessDurationMinutes` parameter to New-AzSnapshotConfig.